Split player implementation over several files
