+++ Sources/Castor/CastPlayer+SkipToDefault.swift
@@ -1,0 +1,44 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import CoreMedia
+
+public extension CastPlayer {
+    /// Returns whether the current player item player can be returned to its default position.
+    ///
+    /// - Returns: `true` if skipping to the default position is possible.
+    ///
+    /// For a livestream supporting DVR this method can be used to check whether the stream is played at the live
+    /// edge or not.
+    func canSkipToDefault() -> Bool {
+        guard let mediaInformation else { return false }
+        switch mediaInformation.streamType {
+        case .live where seekableTimeRange().isValidAndNotEmpty:
+            return time() < seekableTimeRange().end - forwardSkipTime
+        case .live:
+            return false
+        case .buffered:
+            return true
+        default:
+            return false
+        }
+    }
+
+    /// Returns the current item to its default position.
+    ///
+    /// For a livestream supporting DVR the default position corresponds to the live edge.
+    func skipToDefault() {
+        guard let mediaInformation else { return }
+        switch mediaInformation.streamType {
+        case .live:
+            seek(to: seekableTimeRange().end)
+        case .buffered:
+            seek(to: seekableTimeRange().start)
+        default:
+            return
+        }
+    }
+}
