+++ Sources/Castor/CastPlayer+Time.swift
@@ -1,0 +1,39 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+import CoreMedia
+import GoogleCast
+
+public extension CastPlayer {
+    /// Time.
+    func time() -> CMTime {
+        .init(seconds: remoteMediaClient.approximateStreamPosition(), preferredTimescale: CMTimeScale(NSEC_PER_SEC))
+    }
+
+    /// Seekable time range.
+    func seekableTimeRange() -> CMTimeRange {
+        let start = remoteMediaClient.approximateLiveSeekableRangeStart()
+        let end = remoteMediaClient.approximateLiveSeekableRangeEnd()
+        if Self.isValidTimeInterval(start), Self.isValidTimeInterval(end), start != end {
+            return .init(
+                start: .init(seconds: start, preferredTimescale: CMTimeScale(NSEC_PER_SEC)),
+                end: .init(seconds: end, preferredTimescale: CMTimeScale(NSEC_PER_SEC))
+            )
+        }
+        else if let streamDuration = mediaInformation?.streamDuration, Self.isValidTimeInterval(streamDuration), streamDuration != 0 {
+            return .init(start: .zero, end: .init(seconds: streamDuration, preferredTimescale: CMTimeScale(NSEC_PER_SEC)))
+        }
+        else {
+            return .invalid
+        }
+    }
+}
+
+private extension CastPlayer {
+    static func isValidTimeInterval(_ timeInterval: TimeInterval) -> Bool {
+        GCKIsValidTimeInterval(timeInterval) && timeInterval != .infinity
+    }
+}
