+++ Sources/Castor/CastPlayer.swift
@@ -6,37 +6,35 @@
 
 import AVFoundation
 import Combine
-import CoreMedia
 import GoogleCast
-import SwiftUI
 
 /// A cast player.
 public final class CastPlayer: NSObject, ObservableObject {
     let remoteMediaClient: GCKRemoteMediaClient
 
-    @ReceiverState(MediaStatusRecipe.self)
-    private var _mediaStatus
-
-    @MutableReceiverState(ShouldPlayRecipe.self)
-    private var _shouldPlay
-
     @MutableReceiverState(RepeatModeRecipe.self)
     private var _repeatMode
 
-    @MutableReceiverState(PlaybackSpeedRecipe.self)
-    private var _playbackSpeed
-
-    @MutableReceiverState(ActiveTracksRecipe.self)
-    private var _activeTracks
-
-    @MutableReceiverState(TargetSeekTimeRecipe.self)
-    private var _targetSeekTime
-
-    @MutableReceiverState(ItemsRecipe.self)
-    private var _items
-
     @MutableReceiverState(CurrentItemIdRecipe.self)
     private var _currentItemId
+
+    @ReceiverState(MediaStatusRecipe.self)
+    var _mediaStatus
+
+    @MutableReceiverState(ShouldPlayRecipe.self)
+    var _shouldPlay
+
+    @MutableReceiverState(PlaybackSpeedRecipe.self)
+    var _playbackSpeed
+
+    @MutableReceiverState(ActiveTracksRecipe.self)
+    var _activeTracks
+
+    @MutableReceiverState(TargetSeekTimeRecipe.self)
+    var _targetSeekTime
+
+    @MutableReceiverState(ItemsRecipe.self)
+    var _items
 
     /// The current item.
     ///
@@ -51,16 +49,6 @@ public final class CastPlayer: NSObject, ObservableObject {
         }
     }
 
-    /// A Boolean value whether the player should play content when possible.
-    public var shouldPlay: Bool {
-        get {
-            _shouldPlay
-        }
-        set {
-            _shouldPlay = newValue
-        }
-    }
-
     /// The mode with which the player repeats playback of items in its queue.
     public var repeatMode: CastRepeatMode {
         get {
@@ -68,24 +56,6 @@ public final class CastPlayer: NSObject, ObservableObject {
         }
         set {
             _repeatMode = newValue
-        }
-    }
-
-    /// A Boolean value whether the player is active.
-    ///
-    /// Actions performed on an inactive player will usually do nothing. User interfaces should usually disable or hide
-    /// controls when a player is not active.
-    public var isActive: Bool {
-        remoteMediaClient.canMakeRequest()
-    }
-
-    /// The items in the queue.
-    public var items: [CastPlayerItem] {
-        get {
-            _items
-        }
-        set {
-            _items = newValue
         }
     }
 
@@ -99,344 +69,13 @@ public final class CastPlayer: NSObject, ObservableObject {
 
         super.init()
 
+        __repeatMode.bind(to: remoteMediaClient)
+        __currentItemId.bind(to: remoteMediaClient)
         __mediaStatus.bind(to: remoteMediaClient)
         __shouldPlay.bind(to: remoteMediaClient)
-        __repeatMode.bind(to: remoteMediaClient)
         __playbackSpeed.bind(to: remoteMediaClient)
         __activeTracks.bind(to: remoteMediaClient)
         __targetSeekTime.bind(to: remoteMediaClient)
         __items.bind(to: remoteMediaClient)
-        __currentItemId.bind(to: remoteMediaClient)
-    }
-}
-
-public extension CastPlayer {
-    /// Plays.
-    func play() {
-        shouldPlay = true
-    }
-
-    /// Pauses.
-    func pause() {
-        shouldPlay = false
-    }
-
-    /// Toggles between play and pause.
-    func togglePlayPause() {
-        shouldPlay.toggle()
-    }
-
-    /// Stops.
-    func stop() {
-        remoteMediaClient.stop()
-    }
-}
-
-public extension CastPlayer {
-    /// The currently allowed playback speed range.
-    var playbackSpeedRange: ClosedRange<Float> {
-        _mediaStatus?.mediaInformation?.streamType == .buffered ? 0.5...2 : 1...1
-    }
-
-    /// The currently applicable playback speed.
-    var playbackSpeed: Float {
-        get {
-            _playbackSpeed
-        }
-        set {
-            _playbackSpeed = newValue
-        }
-    }
-}
-
-public extension CastPlayer {
-    /// The set of media characteristics for which a media selection is available.
-    var mediaSelectionCharacteristics: Set<AVMediaCharacteristic> {
-        Set(Self.tracks(from: _mediaStatus).compactMap(\.mediaCharacteristic))
-    }
-
-    private static func tracks(from mediaStatus: GCKMediaStatus?) -> [CastMediaTrack] {
-        guard let rawTracks = mediaStatus?.mediaInformation?.mediaTracks else { return [] }
-        return rawTracks.map { .init(rawTrack: $0) }
-    }
-
-    /// Selects a media option for a characteristic.
-    ///
-    /// - Parameters:
-    ///   - mediaOption: The option to select.
-    ///   - characteristic: The characteristic.
-    ///
-    /// You can use `mediaSelectionCharacteristics` to retrieve available characteristics. This method does nothing when
-    /// attempting to set an option that is not supported.
-    func select(mediaOption: CastMediaSelectionOption, for characteristic: AVMediaCharacteristic) {
-        var activeTracks = _activeTracks
-        activeTracks.removeAll { $0.mediaCharacteristic == characteristic }
-        switch mediaOption {
-        case .off:
-            break
-        case let .on(track):
-            activeTracks.append(track)
-        }
-        _activeTracks = activeTracks
-    }
-
-    /// The list of media options associated with a characteristic.
-    ///
-    /// - Parameter characteristic: The characteristic.
-    /// - Returns: The list of options associated with the characteristic.
-    ///
-    /// Use `mediaSelectionCharacteristics` to retrieve available characteristics.
-    func mediaSelectionOptions(for characteristic: AVMediaCharacteristic) -> [CastMediaSelectionOption] {
-        let tracks = Self.tracks(from: _mediaStatus).filter { $0.mediaCharacteristic == characteristic }
-        switch characteristic {
-        case .audible where tracks.count > 1:
-            return tracks.map { .on($0) }
-        case .legible where !tracks.isEmpty:
-            return [.off] + tracks.map { .on($0) }
-        default:
-            return []
-        }
-    }
-
-    /// The currently selected media option for a characteristic.
-    ///
-    /// - Parameter characteristic: The characteristic.
-    /// - Returns: The selected option.
-    ///
-    /// You can use `mediaSelectionCharacteristics` to retrieve available characteristics.
-    func selectedMediaOption(for characteristic: AVMediaCharacteristic) -> CastMediaSelectionOption {
-        let options = mediaSelectionOptions(for: characteristic)
-        let currentOption = currentMediaOption(for: characteristic)
-        return options.contains(currentOption) ? currentOption : .off
-    }
-
-    /// A binding to read and write the current media selection for a characteristic.
-    ///
-    /// - Parameter characteristic: The characteristic.
-    /// - Returns: The binding.
-    func mediaOption(for characteristic: AVMediaCharacteristic) -> Binding<CastMediaSelectionOption> {
-        .init {
-            self.selectedMediaOption(for: characteristic)
-        } set: { newValue in
-            self.select(mediaOption: newValue, for: characteristic)
-        }
-    }
-
-    /// The current media option for a characteristic.
-    ///
-    /// - Parameter characteristic: The characteristic.
-    /// - Returns: The current option.
-    ///
-    /// Unlike `selectedMediaOption(for:)` this method provides the currently applied option. This method can
-    /// be useful if you need to access the actual selection made by `select(mediaOption:for:)` for `.automatic`
-    /// and `.off` options (forced options might be returned where applicable).
-    func currentMediaOption(for characteristic: AVMediaCharacteristic) -> CastMediaSelectionOption {
-        switch characteristic {
-        case .audible, .legible:
-            guard let track = _activeTracks.first(where: { $0.mediaCharacteristic == characteristic }) else { return .off }
-            return .on(track)
-        default:
-            return .off
-        }
-    }
-}
-
-public extension CastPlayer {
-    /// Player state.
-    var state: GCKMediaPlayerState {
-        _mediaStatus?.playerState ?? .unknown
-    }
-
-    /// Media information.
-    var mediaInformation: GCKMediaInformation? {
-        _mediaStatus?.mediaInformation
-    }
-
-    /// Returns if the player is busy.
-    var isBusy: Bool {
-        state == .buffering || state == .loading
-    }
-
-    /// Time.
-    func time() -> CMTime {
-        .init(seconds: remoteMediaClient.approximateStreamPosition(), preferredTimescale: CMTimeScale(NSEC_PER_SEC))
-    }
-
-    /// Seekable time range.
-    func seekableTimeRange() -> CMTimeRange {
-        let start = remoteMediaClient.approximateLiveSeekableRangeStart()
-        let end = remoteMediaClient.approximateLiveSeekableRangeEnd()
-        if Self.isValidTimeInterval(start), Self.isValidTimeInterval(end), start != end {
-            return .init(
-                start: .init(seconds: start, preferredTimescale: CMTimeScale(NSEC_PER_SEC)),
-                end: .init(seconds: end, preferredTimescale: CMTimeScale(NSEC_PER_SEC))
-            )
-        }
-        else if let streamDuration = mediaInformation?.streamDuration, Self.isValidTimeInterval(streamDuration), streamDuration != 0 {
-            return .init(start: .zero, end: .init(seconds: streamDuration, preferredTimescale: CMTimeScale(NSEC_PER_SEC)))
-        }
-        else {
-            return .invalid
-        }
-    }
-}
-
-public extension CastPlayer {
-    /// Performs a seek to a given time.
-    ///
-    /// - Parameter time: The time to reach.
-    func seek(to time: CMTime) {
-        _targetSeekTime = time
-    }
-}
-
-extension CastPlayer {
-    var backwardSkipTime: CMTime {
-        CMTime(seconds: -configuration.backwardSkipInterval, preferredTimescale: 1)
-    }
-
-    var forwardSkipTime: CMTime {
-        CMTime(seconds: configuration.forwardSkipInterval, preferredTimescale: 1)
-    }
-}
-
-public extension CastPlayer {
-    /// Checks whether seeking to a specific time is possible.
-    ///
-    /// - Parameter time: The time to seek to.
-    /// - Returns: `true` if possible.
-    func canSeek(to time: CMTime) -> Bool {
-        let seekableTimeRange = seekableTimeRange()
-        guard seekableTimeRange.isValidAndNotEmpty else { return false }
-        return seekableTimeRange.start <= time && time <= seekableTimeRange.end
-    }
-
-    /// Checks whether skipping backward is possible.
-    ///
-    /// - Returns: `true` if possible.
-    func canSkipBackward() -> Bool {
-        seekableTimeRange().isValidAndNotEmpty
-    }
-
-    /// Checks whether skipping forward is possible.
-    ///
-    /// - Returns: `true` if possible.
-    func canSkipForward() -> Bool {
-        let seekableTimeRange = seekableTimeRange()
-        guard seekableTimeRange.isValidAndNotEmpty else { return false }
-        let currentTime = _targetSeekTime ?? time()
-        return canSeek(to: currentTime + forwardSkipTime)
-    }
-
-    /// Returns whether the current player item player can be returned to its default position.
-    ///
-    /// - Returns: `true` if skipping to the default position is possible.
-    ///
-    /// For a livestream supporting DVR this method can be used to check whether the stream is played at the live
-    /// edge or not.
-    func canSkipToDefault() -> Bool {
-        guard let mediaInformation else { return false }
-        switch mediaInformation.streamType {
-        case .live where seekableTimeRange().isValidAndNotEmpty:
-            return time() < seekableTimeRange().end - forwardSkipTime
-        case .live:
-            return false
-        case .buffered:
-            return true
-        default:
-            return false
-        }
-    }
-
-    /// Checks whether skipping in some direction is possible.
-    ///
-    /// - Returns: `true` if possible.
-    func canSkip(_ skip: CastSkip) -> Bool {
-        switch skip {
-        case .backward:
-            return canSkipBackward()
-        case .forward:
-            return canSkipForward()
-        }
-    }
-}
-
-public extension CastPlayer {
-    /// Skips backward.
-    func skipBackward() {
-        let currentTime = _targetSeekTime ?? time()
-        seek(to: CMTimeClampToRange(currentTime + backwardSkipTime, range: seekableTimeRange()))
-    }
-
-    /// Skips forward.
-    func skipForward() {
-        let currentTime = _targetSeekTime ?? time()
-        seek(to: CMTimeClampToRange(currentTime + forwardSkipTime, range: seekableTimeRange()))
-    }
-
-    /// Skips in a given direction.
-    ///
-    /// - Parameter skip: The skip direction.
-    func skip(_ skip: CastSkip) {
-        switch skip {
-        case .backward:
-            skipBackward()
-        case .forward:
-            skipForward()
-        }
-    }
-
-    /// Returns the current item to its default position.
-    ///
-    /// For a livestream supporting DVR the default position corresponds to the live edge.
-    func skipToDefault() {
-        guard let mediaInformation else { return }
-        switch mediaInformation.streamType {
-        case .live:
-            seek(to: seekableTimeRange().end)
-        case .buffered:
-            seek(to: seekableTimeRange().start)
-        default:
-            return
-        }
-    }
-}
-
-extension CastPlayer {
-    private func pulsePublisher(interval: CMTime) -> AnyPublisher<Void, Never> {
-        Timer.publish(every: interval.seconds, on: .main, in: .common)
-            .autoconnect()
-            .map { _ in }
-            .prepend(())
-            .eraseToAnyPublisher()
-    }
-
-    private func smoothTimePublisher(interval: CMTime) -> AnyPublisher<CMTime, Never> {
-        Publishers.CombineLatest3(
-            $_targetSeekTime,
-            objectWillChange,
-            pulsePublisher(interval: interval)
-        )
-        .weakCapture(self)
-        .map { ($0.0, $1) }
-        .map { targetSeekTime, player in
-            targetSeekTime ?? player.time()
-        }
-        .eraseToAnyPublisher()
-    }
-
-    func timePropertiesPublisher(interval: CMTime) -> AnyPublisher<TimeProperties, Never> {
-        smoothTimePublisher(interval: interval)
-            .weakCapture(self)
-            .map { time, player in
-                TimeProperties(time: time, timeRange: player.seekableTimeRange())
-            }
-            .eraseToAnyPublisher()
-    }
-}
-
-private extension CastPlayer {
-    static func isValidTimeInterval(_ timeInterval: TimeInterval) -> Bool {
-        GCKIsValidTimeInterval(timeInterval) && timeInterval != .infinity
     }
 }
